+++ packages/pluggableWidgets/notifications-native/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Structure mode preview
+
 ## [3.1.1] - 2022-3-9
 ### Fixed
 - We fixed the 'On open' action being triggered twice when tapping a notification while the app was in a background or quit state.